+++ test/argument.test.ts
@@ -16,8 +16,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "npm");
             assert.equal(args[1].getValue(), "install");
-            assert.equal(args[0].getRawValue(), "npm");
-            assert.equal(args[1].getRawValue(), "install");
             assertRange(args[0].getRange(), 0, 4, 0, 7);
             assertRange(args[1].getRange(), 0, 8, 0, 15);
         });
@@ -28,8 +26,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "8000");
             assert.equal(args[1].getValue(), "8001");
-            assert.equal(args[0].getRawValue(), "80\\\n00");
-            assert.equal(args[1].getRawValue(), "8001");
             assertRange(args[0].getRange(), 0, 7, 1, 2);
             assertRange(args[1].getRange(), 1, 3, 1, 7);
         });
@@ -40,8 +36,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "8000");
             assert.equal(args[1].getValue(), "8001");
-            assert.equal(args[0].getRawValue(), "8000");
-            assert.equal(args[1].getRawValue(), "8001");
             assertRange(args[0].getRange(), 0, 7, 0, 11);
             assertRange(args[1].getRange(), 1, 1, 1, 5);
         });
@@ -52,8 +46,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "80");
             assert.equal(args[1].getValue(), "81");
-            assert.equal(args[0].getRawValue(), "80");
-            assert.equal(args[1].getRawValue(), "81");
             assertRange(args[0].getRange(), 0, 7, 0, 9);
             assertRange(args[1].getRange(), 0, 11, 0, 13);
         });
@@ -63,7 +55,6 @@ describe("Argument", () => {
             let args = dockerfile.getInstructions()[0].getArguments();
             assert.equal(args.length, 1);
             assert.equal(args[0].getValue(), "8000");
-            assert.equal(args[0].getRawValue(), "8000");
             assertRange(args[0].getRange(), 0, 9, 0, 13);
         });
 
@@ -75,10 +66,6 @@ describe("Argument", () => {
             assert.equal(args[1].getValue(), "\"a");
             assert.equal(args[2].getValue(), "b\"");
             assert.equal(args[3].getValue(), "]");
-            assert.equal(args[0].getRawValue(), "[");
-            assert.equal(args[1].getRawValue(), "\"a");
-            assert.equal(args[2].getRawValue(), "b\"");
-            assert.equal(args[3].getRawValue(), "]");
             assertRange(args[0].getRange(), 0, 6, 0, 7);
             assertRange(args[1].getRange(), 0, 8, 0, 10);
             assertRange(args[2].getRange(), 0, 12, 0, 14);
@@ -90,7 +77,6 @@ describe("Argument", () => {
             let args = dockerfile.getInstructions()[0].getArguments();
             assert.equal(args.length, 1);
             assert.equal(args[0].getValue(), "a");
-            assert.equal(args[0].getRawValue(), "\\a");
             assertRange(args[0].getRange(), 0, 7, 0, 9);
         });
 
@@ -99,7 +85,6 @@ describe("Argument", () => {
             let args = dockerfile.getInstructions()[0].getArguments();
             assert.equal(args.length, 1);
             assert.equal(args[0].getValue(), "8081");
-            assert.equal(args[0].getRawValue(), "80\\81");
             assertRange(args[0].getRange(), 0, 7, 0, 12);
         });
 
@@ -108,7 +93,6 @@ describe("Argument", () => {
             let args = dockerfile.getInstructions()[0].getArguments();
             assert.equal(args.length, 1);
             assert.equal(args[0].getValue(), "alpine");
-            assert.equal(args[0].getRawValue(), "alpine");
             assertRange(args[0].getRange(), 0, 5, 0, 11);
         });
 
@@ -119,9 +103,6 @@ describe("Argument", () => {
             assert.equal(args[0].getValue(), "alpine");
             assert.equal(args[1].getValue(), "AS");
             assert.equal(args[2].getValue(), "stage");
-            assert.equal(args[0].getRawValue(), "alpine");
-            assert.equal(args[1].getRawValue(), "AS");
-            assert.equal(args[2].getRawValue(), "stage");
             assertRange(args[0].getRange(), 0, 5, 0, 11);
             assertRange(args[1].getRange(), 2, 1, 2, 3);
             assertRange(args[2].getRange(), 2, 4, 2, 9);
@@ -133,8 +114,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "a=a");
             assert.equal(args[1].getValue(), "b");
-            assert.equal(args[0].getRawValue(), "a=a");
-            assert.equal(args[1].getRawValue(), "b");
             assertRange(args[0].getRange(), 0, 4, 0, 7);
             assertRange(args[1].getRange(), 1, 1, 1, 2);
         });
@@ -145,8 +124,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "key=value");
             assert.equal(args[1].getValue(), "key2=value2");
-            assert.equal(args[0].getRawValue(), "key=value");
-            assert.equal(args[1].getRawValue(), "key2=value2");
             assertRange(args[0].getRange(), 0, 4, 0, 13);
             assertRange(args[1].getRange(), 0, 15, 0, 26);
         });
@@ -157,8 +134,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "key=value");
             assert.equal(args[1].getValue(), "key2=value2");
-            assert.equal(args[0].getRawValue(), "key=value");
-            assert.equal(args[1].getRawValue(), "key2=value2");
             assertRange(args[0].getRange(), 0, 6, 0, 15);
             assertRange(args[1].getRange(), 0, 17, 0, 28);
         });
@@ -171,8 +146,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "echo");
             assert.equal(args[1].getValue(), "value");
-            assert.equal(args[0].getRawValue(), "echo");
-            assert.equal(args[1].getRawValue(), "$var");
             assertRange(args[0].getRange(), 1, 4, 1, 8);
             assertRange(args[1].getRange(), 1, 9, 1, 13);
         });
@@ -183,8 +156,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "echo");
             assert.equal(args[1].getValue(), "value$var2");
-            assert.equal(args[0].getRawValue(), "echo");
-            assert.equal(args[1].getRawValue(), "$var$var2");
             assertRange(args[0].getRange(), 1, 4, 1, 8);
             assertRange(args[1].getRange(), 1, 9, 1, 18);
         });
@@ -195,8 +166,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "echo");
             assert.equal(args[1].getValue(), "value$var2");
-            assert.equal(args[0].getRawValue(), "echo");
-            assert.equal(args[1].getRawValue(), "$var$var2");
             assertRange(args[0].getRange(), 1, 4, 1, 8);
             assertRange(args[1].getRange(), 1, 9, 1, 18);
         });
@@ -207,8 +176,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "echo");
             assert.equal(args[1].getValue(), "$var$var2");
-            assert.equal(args[0].getRawValue(), "echo");
-            assert.equal(args[1].getRawValue(), "$var$var2");
             assertRange(args[0].getRange(), 0, 4, 0, 8);
             assertRange(args[1].getRange(), 0, 9, 0, 18);
         });
@@ -219,8 +186,6 @@ describe("Argument", () => {
             assert.equal(args.length, 2);
             assert.equal(args[0].getValue(), "var=value");
             assert.equal(args[1].getValue(), "var2=value2");
-            assert.equal(args[0].getRawValue(), "var=value");
-            assert.equal(args[1].getRawValue(), "var2=value2");
             assertRange(args[0].getRange(), 0, 4, 0, 13);
             assertRange(args[1].getRange(), 1, 1, 1, 12);
         });