+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Removed
+- `Argument`'s `getRawValue()` function has been removed [(#10)](https://github.com/rcjsuen/dockerfile-ast/issues/10)
+```TypeScript
+// this convenience function has been removed
+let rawValue = argument.getRawValue();
+// to retrieve the identical value, use the following code instead
+import { TextDocument } from 'vscode-languageserver-types';
+let document = TextDocument.create(uri, languageId, version, buffer);
+let range = argument.getRange();
+let rawValue = buffer.substring(document.offsetAt(range.start), document.offsetAt(range.end));
+```
 
 ## 0.0.1 - 2017-12-20
 ### Added